KAN-647 enhance: 메시징 디렉토리 infrastructure 디렉토리로 이동
