+++ src/main/java/com/yeonieum/productservice/domain/product/dto/customerservice/TimesaleRequestForCustomer.java
@@ -3,9 +3,8 @@ package com.yeonieum.productservice.domain.product.dto.customerservice;
 import com.yeonieum.productservice.domain.product.entity.Product;
 import com.yeonieum.productservice.domain.product.entity.ProductTimesale;
 import com.yeonieum.productservice.domain.product.entity.ServiceStatus;
-import com.yeonieum.productservice.global.enums.ActiveStatus;
 import com.yeonieum.productservice.global.enums.ServiceStatusCode;
-import com.yeonieum.productservice.messaging.message.TimesaleEventMessage;
+import com.yeonieum.productservice.infrastructure.messaging.message.TimesaleEventMessage;
 import jakarta.validation.constraints.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
